requirements.txt
pandas
numpy
tensorflow-cpu
keras
scikit-learn
